requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
filelock==3.12.4
ftfy==6.1.1
huggingface-hub==0.17.3
idna==3.4
numpy==1.24.4
packaging==23.2
Pillow==10.1.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
tokenizers==0.14.1
tqdm==4.66.1
transformers~=4.34.1
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.8
opencv-python~=4.8.1.78
albumentations~=1.3.1
matplotlib~=3.7.3
timm~=0.9.7
av~=10.0.0
textdistance~=4.6.0
torchserve~=0.9.0
torch-model-archiver~=0.9.0
torch-workflow-archiver
natsort~=8.4.0
faiss-cpu~=1.7.1
nvgpu~=0.10.0
setuptools~=68.2.0
pytube~=15.0.0
python-dateutil~=2.8.2
nltk~=3.8.1
coloredlogs~=15.0.1
colorama~=0.4.6
sentence-transformers~=2.2.2
